+++ source/fortran/evaluators.f90
@@ -54,8 +54,8 @@ contains
 !     !  Arguments
 
 !     integer, intent(in) :: n
-!     real(dp), dimension(n), intent(inout) :: xc
-!     real(dp), dimension(n), intent(out) :: rc
+!     real(8), dimension(n), intent(inout) :: xc
+!     real(8), dimension(n), intent(out) :: rc
 !     integer, intent(inout) :: iflag
 
 !     !  Local variables
@@ -108,12 +108,12 @@ contains
 
     !  Arguments
 
-    real(dp), intent(out) :: fc
+    real(8), intent(out) :: fc
 !    real(c_double), intent(out) :: fc
     !  Local variables
 
     integer :: iab
-    real(dp) :: sgn
+    real(8) :: sgn
 
 !        write(*,*) 'Figure of merit 2 (fusion power / input power) is not used.'
     ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
