+++ source/fortran/ife.f90
@@ -171,7 +171,7 @@ contains
 
     !  Local variables
 
-    real(dp) :: aaion,bmax,dpp,dtheta,emitt,etai,lf,phi,qion, &
+    real(8) :: aaion,bmax,dpp,dtheta,emitt,etai,lf,phi,qion, &
          sang,sigma,sigma0,tauf,theta,vi
     integer :: nbeams
 
@@ -335,21 +335,21 @@ contains
 
     !  Arguments
 
-    real(dp), intent(in) :: aaion,bmax,dpp,dtheta,edrive,emitt,etai,lf,qion, &
+    real(8), intent(in) :: aaion,bmax,dpp,dtheta,edrive,emitt,etai,lf,qion, &
          sigma,sigma0,tauf,theta,vi
     integer, intent(in) :: nbeams
-    real(dp), intent(out) :: gain,etadrv
+    real(8), intent(out) :: gain,etadrv
 
     !  Local variables
 
-    real(dp), parameter :: c2 = 8.98755178737D16
+    real(8), parameter :: c2 = 8.98755178737D16
     integer, parameter :: isimp = 1  !  Switch for simple model (1=yes)
 
-    real(dp) :: ci,de,dgap,dlcore,drcore,e,eomc2,fins,floss, &
+    real(8) :: ci,de,dgap,dlcore,drcore,e,eomc2,fins,floss, &
          ibf,ibfo,ibi,ibpc,lfocus,lpf,lpfo,lpi,lppc,lq,phif,phifo, &
          phii,phipc,rion,rs,rs1,rs2,rs3,rs4,sig,sig0,taufo,taui, &
          taupc,tbrad,vf,vfo,vpc,vs,vscore,vshe,vsle,vspc,xhe,xle,xpc
-    real(dp), dimension(10) :: eve,gve
+    real(8), dimension(10) :: eve,gve
     integer :: i,ie,j,nche,ncle,ncpc,nqarrs,nqle,nqpche,nquads
 
     ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
@@ -706,16 +706,16 @@ contains
 
       implicit none
 
-      real(dp) :: cbeam
+      real(8) :: cbeam
 
       !  Arguments
 
-      real(dp), intent(in) :: etai,bmax,aaion,qion,sigma,sigma0, &
+      real(8), intent(in) :: etai,bmax,aaion,qion,sigma,sigma0, &
            emitt,vi,eomc2
 
       !  Local variables
 
-      real(dp) :: bbe
+      real(8) :: bbe
 
       ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
@@ -752,15 +752,15 @@ contains
 
       implicit none
 
-      real(dp) :: betgam
+      real(8) :: betgam
 
       !  Arguments
 
-      real(dp), intent(in) :: aaion,qion,v
+      real(8), intent(in) :: aaion,qion,v
 
       !  Local variables
 
-      real(dp), parameter :: c2 = 8.98755178737D16
+      real(8), parameter :: c2 = 8.98755178737D16
 
       ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
@@ -794,13 +794,13 @@ contains
 
     !  Arguments
 
-    real(dp), intent(in) :: edrive
-    real(dp), intent(out) :: etadrv,gain
+    real(8), intent(in) :: edrive
+    real(8), intent(out) :: etadrv,gain
 
     !  Local variables
 
-    real(dp) :: e,de
-    real(dp), dimension(10) :: gve,eve
+    real(8) :: e,de
+    real(8), dimension(10) :: gve,eve
     integer :: ie
 
     ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
@@ -886,13 +886,13 @@ contains
 
     !  Arguments
 
-    real(dp), intent(in) :: edrive
-    real(dp), dimension(10), intent(in) :: etave,gainve
-    real(dp), intent(out) :: etadrv,gain
+    real(8), intent(in) :: edrive
+    real(8), dimension(10), intent(in) :: etave,gainve
+    real(8), intent(out) :: etadrv,gain
 
     !  Local variables
 
-    real(dp) :: de,e
+    real(8) :: de,e
     integer :: ie
 
     ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
@@ -1192,7 +1192,7 @@ contains
      !  Local variables
 
      integer :: i,j
-     real(dp), save :: g, vel, acurt, mdot, phi, sang, li_frac
+     real(8), save :: g, vel, acurt, mdot, phi, sang, li_frac
 
      ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
@@ -1542,8 +1542,8 @@ contains
 
       !  Local variables
 
-      real(dp), parameter :: third  = 1.0D0/3.0D0
-      real(dp) :: chcylh,ddz,dvol
+      real(8), parameter :: third  = 1.0D0/3.0D0
+      real(8) :: chcylh,ddz,dvol
       integer :: i,j
 
       ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
@@ -1686,8 +1686,8 @@ contains
 
       !  Local variables
 
-      real(dp), parameter :: third = 1.0D0/3.0D0
-      real(dp) :: chcylh,ddz,dvol
+      real(8), parameter :: third = 1.0D0/3.0D0
+      real(8) :: chcylh,ddz,dvol
       integer :: i,j
 
       ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
@@ -1874,8 +1874,8 @@ contains
 
     !  Local variables
 
-    real(dp) :: den,life
-    real(dp), dimension(0:maxmat) :: matden
+    real(8) :: den,life
+    real(8), dimension(0:maxmat) :: matden
     integer :: i,j
 
     ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
@@ -2010,7 +2010,7 @@ contains
 
     !  Local variables
 
-    real(dp) :: pdrvmw
+    real(8) :: pdrvmw
 
     ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
@@ -2096,7 +2096,7 @@ contains
 
     !  Local variables
 
-    real(dp), save :: basemw,pmwpm2
+    real(8), save :: basemw,pmwpm2
 
     ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
@@ -2308,9 +2308,9 @@ contains
 
     !  Local variables
 
-    real(dp) :: cran,dcl,dcw,fac2,fac3,hcl,hcw,hrbi,rbh,rbl,rbw, &
+    real(8) :: cran,dcl,dcw,fac2,fac3,hcl,hcw,hrbi,rbh,rbl,rbw, &
          rmbh,rmbl,rmbw,rwl,rww,shh,tch,tcl,tcw,wgts,wsa
-    real(dp), save :: cryv,elev,rbv,rmbv,vrci,wsv
+    real(8), save :: cryv,elev,rbv,rmbv,vrci,wsv
 
     ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
