+++ source/fortran/pulse.f90
@@ -95,7 +95,7 @@ contains
 
     !  Local variables
 
-    real(dp) :: albusa,ioht1,ioht2,ipdot,loh,m,pfbusl,r,rho,v
+    real(8) :: albusa,ioht1,ioht2,ipdot,loh,m,pfbusl,r,rho,v
 
     ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
@@ -189,7 +189,7 @@ contains
 
     !  Local variables
 
-    real(dp) :: tb,vburn,vsmax,vssoft
+    real(8) :: tb,vburn,vsmax,vssoft
 
     ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
