+++ source/fortran/reinke_module.f90
@@ -7,7 +7,7 @@ module reinke_module
   !  Module-level variables
 
   !integer ::
-  real(dp) :: vcritx
+  real(8) :: vcritx
 
 contains
 
@@ -51,16 +51,16 @@ contains
     use impurity_radiation_module, only: nimp, imp_label
     
     implicit none
-    real(dp) :: reinke_fzmin
-    real(dp) :: bt, flh, qstar, rmajor, eps, fsep, fgw, kappa
-    real(dp) :: lhat, netau, tesep, ml_div, sum_fZ_ml_other, ml_z, lz
-    real(dp), dimension(14) :: impurity_arr_frac
-    real(dp), dimension(14) :: impurity_enrichment
+    real(8) :: reinke_fzmin
+    real(8) :: bt, flh, qstar, rmajor, eps, fsep, fgw, kappa
+    real(8) :: lhat, netau, tesep, ml_div, sum_fZ_ml_other, ml_z, lz
+    real(8), dimension(14) :: impurity_arr_frac
+    real(8), dimension(14) :: impurity_enrichment
     integer(kind=4) :: impvardiv
 
     integer(kind=4), parameter :: N = 100
     integer(kind=4) :: i, j
-    real(dp) :: binWidth, te
+    real(8) :: binWidth, te
 
     binWidth = tesep / N
     ! mL =1/tesep * \int_0^tesep L\(T) sqrt(T) dT using trapezoidal rule
@@ -154,9 +154,9 @@ contains
     !! M.L. Reinke 2017 Nucl. Fusion 57 034004
 
     implicit none
-    real(dp) :: reinke_tsep
-    real(dp) :: bt, flh, qstar, rmajor, eps, fgw, kappa, lhat
-    real(dp), parameter :: kappa_0 = 2D3 !Stangeby W/m/eV^(7/2)
+    real(8) :: reinke_tsep
+    real(8) :: bt, flh, qstar, rmajor, eps, fgw, kappa, lhat
+    real(8), parameter :: kappa_0 = 2D3 !Stangeby W/m/eV^(7/2)
 
     reinke_tsep = bt**0.72 * flh**0.29 * fgw**0.21 * qstar**0.08 * rmajor**0.33
     !reinke_tsep = bt**0.72 * flh**0.2857 * fgw**0.2057 * qstar**0.08 * rmajor**0.3314
@@ -175,11 +175,11 @@ contains
     use impurity_radiation_module, only: imp_dat, imp_label
     implicit none
 
-    real(dp) :: testResult_fZ_DEMOBASE, testResult_fZ_ASDEXBASE, testInput_tsep
+    real(8) :: testResult_fZ_DEMOBASE, testResult_fZ_ASDEXBASE, testInput_tsep
     integer :: i, j
-    real(dp), parameter :: test_Bt = 5.8547
+    real(8), parameter :: test_Bt = 5.8547
     type(imp_dat),  dimension(14), save :: test_imp_arr
-    real(dp), dimension(14) :: impurity_enrichment
+    real(8), dimension(14) :: impurity_enrichment
 
     do i=1,14
        test_imp_arr(i)%frac = 0.0d0
