+++ source/fortran/scan.f90
@@ -98,10 +98,10 @@ module scan_module
   integer :: nsweep_2
   !! nsweep_2 /3/ : switch denoting quantity to scan for 2D scan:
 
-  real(dp), dimension(ipnscns) :: sweep
+  real(8), dimension(ipnscns) :: sweep
   !! sweep(ipnscns) /../: actual values to use in scan
 
-  real(dp), dimension(ipnscns) :: sweep_2
+  real(8), dimension(ipnscns) :: sweep_2
   !! sweep_2(ipnscns) /../: actual values to use in 2D scan
 
   ! Vars in subroutines scan_1d and scan_2d requiring re-initialising before 
@@ -185,7 +185,7 @@ contains
 
     integer, intent(in) :: iscan
     integer, intent(in) :: ifail
-    real(dp), dimension(:,:), intent(out) :: outvar
+    real(8), dimension(:,:), intent(out) :: outvar
 
     ! Turn off error reporting (until next output)
     errors_on = .false.
@@ -288,7 +288,7 @@ contains
     implicit none
 
     integer, intent(inout) :: iscan
-    real(dp), dimension(:,:), intent(in) :: outvar
+    real(8), dimension(:,:), intent(in) :: outvar
     
     character(len=48) :: tlabel
     integer :: ivar
@@ -473,8 +473,8 @@ contains
     integer, intent(in) :: iscan_1
     integer, intent(in) :: iscan_R
     integer, intent(in) :: iscan
-    real(dp), dimension(:,:), intent(out) :: outvar
-    real(dp), dimension(:), intent(out) :: sweep_1_vals, sweep_2_vals
+    real(8), dimension(:,:), intent(out) :: outvar
+    real(8), dimension(:), intent(out) :: sweep_1_vals, sweep_2_vals
 
     call scan_1d_store_output(iscan, ifail, outvar)
 
@@ -488,8 +488,8 @@ contains
     implicit none
 
     integer, intent(inout) :: iscan
-    real(dp), dimension(:,:), intent(in) :: outvar
-    real(dp), dimension(:), intent(in) :: sweep_1_vals, sweep_2_vals
+    real(8), dimension(:,:), intent(in) :: outvar
+    real(8), dimension(:), intent(in) :: sweep_1_vals, sweep_2_vals
 
     integer :: ivar
     character(len=48) :: tlabel
@@ -621,7 +621,7 @@ contains
 
     ! Arguments
     integer, intent(in) :: nwp, iscn
-    real(dp), intent(in), dimension(:) :: swp
+    real(8), intent(in), dimension(:) :: swp
     character(len=25), intent(out) :: vlab, xlab
 
     select case (nwp)
@@ -846,8 +846,8 @@ contains
 
   !  Local variables
   integer :: ii,inn,iflag
-  real(dp) :: summ,xcval,xmaxx,xminn,f,xnorm
-  real(dp), dimension(ipeqns) :: con1, con2, err
+  real(8) :: summ,xcval,xmaxx,xminn,f,xnorm
+  real(8), dimension(ipeqns) :: con1, con2, err
   character(len=1), dimension(ipeqns) :: sym
   character(len=10), dimension(ipeqns) :: lab
   character(len=30) :: strfom
